Mostly compiling now

+++ ShadowEngine/src/ShadowRenderer/GraphicsContext.cpp
@@ -16,7 +16,7 @@ namespace ShadowEngine::Rendering {
 			case RendererAPI::API::None:    SH_CORE_ASSERT(false, "RendererAPI::None is currently not supported!"); return nullptr;
 			//case RendererAPI::API::OpenGL:  return std::make_shared<OpenGL::OpenGLContext>(window);
 			case RendererAPI::API::D3D12: return std::make_shared <D3D12::D3D12Context>(window);
-			default: SH_CORE_CRITICAL("Unknown RendererAPI: {0} !", RendererAPI::GetAPI()); return nullptr;
+			default: SH_CORE_CRITICAL("Unknown RendererAPI: {0} !", (int)RendererAPI::GetAPI()); return nullptr;
 		}
 	}
